+++ src/yafowil/tests/test_common.py
@@ -3191,6 +3191,35 @@ class TestCommon(YafowilTestCase):
             'name="MYFILE" type="file" />'
         ))
 
+        # File actions vocabulary
+        widget = factory(
+            'file',
+            name='MYFILE',
+            value={
+                'file': StringIO('321')
+            },
+            props={
+                'vocabulary': [
+                    ('keep', 'Keep Existing file'),
+                    ('replace', 'Replace existing file')
+                ]
+            })
+        self.checkOutput("""
+        <div>
+          <input id="input-MYFILE" name="MYFILE" type="file"/>
+          <div id="radio-MYFILE-keep">
+            <input checked="checked" id="input-MYFILE-keep"
+                   name="MYFILE-action" type="radio" value="keep"/>
+            <span>Keep Existing file</span>
+          </div>
+          <div id="radio-MYFILE-replace">
+            <input id="input-MYFILE-replace" name="MYFILE-action"
+                   type="radio" value="replace"/>
+            <span>Replace existing file</span>
+          </div>
+        </div>
+        """, wrapped_fxml(widget()))
+
         # Mimetype extractor
         widget = factory(
             'file',